+++ balance_car/Two_balance_car/Core/Inc/stm32f1xx_it.h
@@ -1,4 +1,4 @@
-/* USER CODE BEGIN Header */
+﻿/* USER CODE BEGIN Header */
 /**
   ******************************************************************************
   * @file    stm32f1xx_it.h
@@ -55,6 +55,7 @@ void SVC_Handler(void);
 void DebugMon_Handler(void);
 void PendSV_Handler(void);
 void SysTick_Handler(void);
+void TIM1_UP_IRQHandler(void);
 void USART1_IRQHandler(void);
 /* USER CODE BEGIN EFP */
 
@@ -65,3 +66,4 @@ void USART1_IRQHandler(void);
 #endif
 
 #endif /* __STM32F1xx_IT_H */
+